+++ CHANGELOG.md
@@ -11,3 +11,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Basic Notecard driver support over I2C and UART
 - Two basic samples, one for the each of the communication busses, that request
   version string from notecard.
+- Interrupt support for `ATTN_P` pin.